fixed csv to download all races instead of just the first
